requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
vnstock==0.1.9
torch==2.5.0
transformers==4.48.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1 